refactor: test class -> main class
