+++ app/src/main/java/com/joonhyeok/app/concert/domain/Seat.java
