+++ app/src/main/java/com/joonhyeok/app/queue/appication/dto/QueueQuery.java
