+++ app/src/main/java/com/joonhyeok/app/queue/appication/dto/QueueQueryETAResult.java
