+++ app/src/main/java/com/joonhyeok/app/queue/domain/Queue.java
