+++ app/src/main/java/com/joonhyeok/app/queue/domain/QueueStatus.java
