+++ app/src/main/java/com/joonhyeok/app/queue/presentation/QueueController.java
@@ -1,0 +1,4 @@
+package com.joonhyeok.app.queue.presentation;
+
+public class QueueController{
+}
