+++ app/src/main/java/com/joonhyeok/app/reservation/domain/ReservationPolicy.java
@@ -1,7 +1,6 @@
 package com.joonhyeok.app.reservation.domain;
 
 import com.joonhyeok.app.concert.domain.Concert;
-import com.joonhyeok.app.concert.domain.PerformanceDate;
 import com.joonhyeok.app.concert.domain.Seat;
 
 public interface ReservationPolicy {
