+++ app/src/main/java/com/joonhyeok/app/reservation/domain/ReservationStatus.java
