+++ app/src/main/java/com/joonhyeok/app/user/UserJpaRepository.java
@@ -1,6 +1,5 @@
 package com.joonhyeok.app.user;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.Repository;
 
 public interface UserJpaRepository extends UserRepository, Repository<User, Long> {
