+++ app/src/test/java/com/joonhyeok/app/AppApplicationTests.java
@@ -1,0 +1,13 @@
+package com.joonhyeok.app;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AppApplicationTests {
+
+//	@Test
+//	void contextLoads() {
+//	}
+
+}
