+++ app/src/test/java/com/joonhyeok/app/concert/ConcertMemoryRepository.java
@@ -1,4 +1,4 @@
-package com.joonhyeok.app.concert.infra;
+package com.joonhyeok.app.concert;
 
 import com.joonhyeok.app.concert.domain.Concert;
 import com.joonhyeok.app.concert.domain.ConcertRepository;
