+++ app/src/test/java/com/joonhyeok/app/concert/SeatMemoryRepository.java
@@ -1,4 +1,4 @@
-package com.joonhyeok.app.concert.infra;
+package com.joonhyeok.app.concert;
 
 import com.joonhyeok.app.concert.domain.Seat;
 import com.joonhyeok.app.concert.domain.SeatRepository;
